clean code in some parts

+++ app/Http/Controllers/MatakuliahController.php
@@ -5,25 +5,23 @@ namespace App\Http\Controllers;
 use App\Models\Matakuliah;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use App\Helpers\ResponseHelper;
 
 class MatakuliahController extends Controller
 {
     public function index()
     {
         $matakuliah = Matakuliah::all();
-        return response()->json([
-            'message' => 'Mata Kuliah berhasil ditampilkan!',
-            'data' => $matakuliah
-        ]);
+        return ResponseHelper::success('Mata Kuliah berhasil ditampilkan!', $matakuliah);
     }
 
     public function show($id)
     {
-        $matakuliah = Matakuliah::findOrFail($id);
-        return response()->json([
-            'message' => 'Mata Kuliah berhasil ditampilkan!',
-            'data' => $matakuliah
-        ]);
+        $matakuliah = Matakuliah::find($id);
+        if (!$matakuliah) {
+                return ResponseHelper::error('Mata Kuliah tidak ditemukan!', 404);
+        }
+        return ResponseHelper::success('Mata Kuliah berhasil ditampilkan!', $matakuliah);
     }
 
     public function store(Request $request)
@@ -39,15 +37,15 @@ class MatakuliahController extends Controller
             'kode_mata_kuliah' => $request->kode_mata_kuliah,
             'sks' => $request->sks,
         ]);
-        return response()->json([
-            'message' => 'Mata Kuliah berhasil ditambahkan!',
-            'data' => $matakuliah
-        ]);
+        return ResponseHelper::success('Mata Kuliah berhasil ditambahkan!', $matakuliah);
     }
 
     public function update(Request $request, $id)
     {
-        $matakuliah = Matakuliah::findOrFail($id);
+        $matakuliah = Matakuliah::find($id);
+        if (!$matakuliah) {
+            return ResponseHelper::error('Mata Kuliah tidak ditemukan!', 404);
+        }
         $request->validate([
             'nama_mata_kuliah' => 'sometimes|string|max:100',
             'kode_mata_kuliah' => 'sometimes|string|max:10|unique:mata_kuliah,kode_mata_kuliah',
@@ -56,22 +54,19 @@ class MatakuliahController extends Controller
 
         $matakuliah->update($request->only('nama_mata_kuliah', 'kode_mata_kuliah', 'sks'));
     
-        return response()->json([
-            'message' => 'Mata Kuliah berhasil diubah!',
-            'data' => $matakuliah
-        ]);
+        return ResponseHelper::success('Mata Kuliah berhasil diubah!', $matakuliah);
     }
 
     public function destroy($id)
     {
-        $matakuliah = Matakuliah::findOrFail($id);
+        $matakuliah = Matakuliah::find($id);
         $matakuliah->mahasiswa()->detach();
+        if (!$matakuliah) {
+            return ResponseHelper::error('Mata Kuliah tidak ditemukan!', 404);
+        }
     
         $matakuliah->delete();
-        return response()->json([
-            'message' => 'Mata Kuliah berhasil dihapus!',
-            'data' => $matakuliah
-        ]);
+        return ResponseHelper::success('Mata Kuliah berhasil dihapus!', $matakuliah);
     }
     
 }
